+++ sampledb/frontend/users/authentication.py
@@ -4,6 +4,7 @@
 """
 
 import datetime
+import json
 import typing
 
 import flask
@@ -11,9 +12,10 @@ import flask_login
 from flask_babel import _, lazy_gettext, refresh
 
 from .. import frontend
-from ...logic.authentication import login, get_active_two_factor_authentication_methods
+from ...logic.authentication import login, get_active_two_factor_authentication_methods, get_all_fido2_passkey_credentials, get_user_id_for_fido2_passkey_credential_id, get_webauthn_server
 from ...logic.users import get_user, User
-from ...frontend.users_forms import SigninForm, SignoutForm
+from ..users_forms import SigninForm, SignoutForm
+from .forms import WebAuthnLoginForm
 from ... import login_manager
 from ...utils import FlaskResponseT
 
@@ -66,6 +68,10 @@ def _redirect_to_next_url() -> FlaskResponseT:
 
 def _sign_in_impl(is_for_refresh: bool) -> FlaskResponseT:
     form = SigninForm()
+    passkey_form = WebAuthnLoginForm()
+    all_credentials = get_all_fido2_passkey_credentials()
+    server = get_webauthn_server()
+
     has_errors = False
     if form.validate_on_submit():
         username = form.username.data
@@ -107,16 +113,45 @@ def _sign_in_impl(is_for_refresh: bool) -> FlaskResponseT:
                 two_factor_authentication_method = two_factor_authentication_methods[0]
                 if two_factor_authentication_method.data.get('type') == 'totp':
                     return flask.redirect(flask.url_for('.confirm_totp_two_factor_authentication', method_id=two_factor_authentication_method.id))
+                if two_factor_authentication_method.data.get('type') == 'fido2_passkey':
+                    return flask.redirect(flask.url_for('.confirm_fido2_passkey_two_factor_authentication', method_id=two_factor_authentication_method.id))
                 del flask.session['confirm_data']
                 return flask.render_template('two_factor_authentication/unsupported_method.html')
             return flask.render_template(
                 'two_factor_authentication/pick.html',
-                methods=two_factor_authentication_methods
+                methods=two_factor_authentication_methods,
             )
         has_errors = True
     elif form.errors:
         has_errors = True
-    return flask.render_template('sign_in.html', form=form, is_for_refresh=is_for_refresh, has_errors=has_errors)
+    if passkey_form.validate_on_submit() and flask.current_app.config['ENABLE_FIDO2_PASSKEY_AUTHENTICATION']:
+        try:
+            credential = server.authenticate_complete(
+                flask.session.pop("webauthn_state"),
+                credentials=all_credentials,
+                response=json.loads(passkey_form.assertion.data),
+            )
+            user_id = get_user_id_for_fido2_passkey_credential_id(credential.credential_id)
+        except Exception:
+            user_id = None
+        if user_id is not None:
+            user = get_user(user_id)
+            return complete_sign_in(user, is_for_refresh, False)
+        flask.flash(_('No user was found for this passkey.'), 'error')
+        return flask.redirect(flask.url_for('frontend.sign_in'))
+
+    options, state = server.authenticate_begin(
+        credentials=all_credentials
+    )
+    flask.session["webauthn_state"] = state
+    return flask.render_template(
+        'sign_in.html',
+        form=form,
+        is_for_refresh=is_for_refresh,
+        has_errors=has_errors,
+        passkey_form=passkey_form,
+        options=options
+    )
 
 
 def complete_sign_in(user: typing.Optional[User], is_for_refresh: bool, remember_me: bool) -> FlaskResponseT: